requirements.txt
requests>=2.31
astral>=3.2
python-dotenv>=1.0
APScheduler>=3.10,<4.0
holidays>=0.46
tzdata>=2024.1
argparse-manpage>=4.5
pytest>=7.4
